postgres working locally

+++ src/main/java/com/codurance/principle12/factories/CardFactory.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 public class CardFactory {
 
   public Card create(NewCardRequestObject requestObject) {
-    int id = CardIDGenerator.nextID();
+    Long id = CardIDGenerator.nextID();
     return new Card(id, requestObject.getText(), requestObject.getColumnId(), requestObject.getUserName());
   }
 }
